worker-prev/requirements.txt
aio-pika==9.5.4
requests
moviepy
openai==1.58.1
whisper
elevenlabs===1.50.3
pika==1.3.2
python-dotenv
ffmpy
jsonschema
aiohttp
redis
ffmpeg-python